% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agesex.R
\name{agesex}
\alias{agesex}
\title{Create population grid for age-sex groups}
\usage{
agesex(agesex_select, pop_raster, agesex_regions, agesex_table)
}
\arguments{
\item{agesex_select}{character vector. Vector of age-sex groups to include in the population estimate (e.g. c("m0","m1","f0","f1") = children under five; "f0" = females less than 1 year, "f1" = females 1 to 4 years, "f5" = females 5 to 9 years, etc.)}

\item{pop_raster}{raster. Gridded population estimates.}

\item{agesex_regions}{raster. Raster of region IDs for age-sex information.}

\item{agesex_table}{data.frame. Proportion of the population in each age-sex group in each region. The first column must contain region IDs that correspond to "agesex_regions". The remaining columns are named in accordance with "agesex_select".}
}
\value{
raster. Gridded population estimates.
}
\description{
Create gridded population estimates for specific age-sex groups by disaggregating population totals using gridded age-sex proportions.
}
