% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agesexLookup.R
\name{agesexLookup}
\alias{agesexLookup}
\title{Format shiny age-group selection for peanutButter}
\usage{
agesexLookup(male, female, male_select, female_select)
}
\arguments{
\item{male}{logical.}

\item{female}{logical.}

\item{male_select}{character vector.}

\item{female_select}{character vector.}
}
\description{
Format shiny age-group selection for peanutButter
}
