% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{aggregator}
\alias{aggregator}
\title{Aggregate population estimates from buildings}
\usage{
aggregator(
  buildings,
  urban,
  people_urb = 5,
  units_urb = 1,
  residential_urb = 0.5,
  people_rur = 5,
  residential_rur = 0.5,
  units_rur = 1
)
}
\arguments{
\item{buildings}{raster. A "raster" object with building counts per pixel}

\item{urban}{raster. A "raster" object with binary map of urban areas (i.e. 0 and 1)}

\item{people_urb}{numeric. Average number of people per housing unit in urban areas}

\item{units_urb}{numeric. Average number of housing units per building in urban areas}

\item{residential_urb}{numeric. Probability of residential building in urban areas}

\item{people_rur}{numeric. Average number of people per housing unit in rural areas}

\item{residential_rur}{numeric. Probability of residential building in rural areas}

\item{units_rur}{numeric. Average number of housing units per building in rural areas}
}
\description{
Create population raster using the peanut butter method with building footprints.
}
