% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{demographic}
\alias{demographic}
\title{Create population grid for age-sex groups}
\usage{
demographic(population, group_select, regions, proportions)
}
\arguments{
\item{population}{raster. Gridded population estimates.}

\item{group_select}{character vector. Vector of age-sex groups to include in the population estimate (i.e. "f0" = females less than 1 year, "f1" = females 1 to 4 years, "f5" = females 5 to 9 years, etc.; c("m0","m1","f0","f1") = children under five)}

\item{regions}{raster. Raster of region IDs for age-sex information.}

\item{proportions}{data.frame. Proportion of the population in each age-sex group in each region. The first column must contain region IDs that correspond to "regions". The remaining columns are named in accordance with "group_select".}
}
\value{
raster. Gridded population estimates.
}
\description{
Create gridded population estimates for specific age-sex groups by disaggregating population totals using gridded age-sex proportions.
}
