% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregator.R
\name{disaggregator}
\alias{disaggregator}
\title{Disaggregate population totals from polygons}
\usage{
disaggregator(feature, buildings, popcol = NULL)
}
\arguments{
\item{feature}{sf. An "sf" object with feature geometries (POLYGONS or MULTIPOLYGONS). The first column should contain population totals for each polygon (class "numeric").}

\item{buildings}{raster. A "raster" object with counts of buildings per pixel.}

\item{popcol}{character. The column name from "feature" that contains the population totals. If NULL, the first column will be used.}
}
\value{
Gridded population estimates as a "raster" object.
}
\description{
Create gridded population estimates by disaggregating population totals from polygon geometries using building footprints.
}
