% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadData}
\alias{downloadData}
\title{Download source data from WOPR}
\usage{
downloadData(version = "v1.0", wopr_dir = "wopr", maxsize = 100)
}
\arguments{
\item{version}{Version of the buildings data sets to use}

\item{wopr_dir}{Directory where downloads should be saved}

\item{maxsize}{Maximum file size (MB) allowed to download without notification}
}
\value{
Files are downloaded to local disk
}
\description{
Download peanutButter source data from the WorldPop Open Population Repository
}
