% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregator.R
\name{disaggregator}
\alias{disaggregator}
\title{Disaggregate population totals from polygons}
\usage{
disaggregator(feature, buildings, popcol = NULL)
}
\arguments{
\item{feature}{An "sf" object with feature geometries (POLYGONS or MULTIPOLYGONS). The first column should contain population totals for each polygon (class "numeric").}

\item{buildings}{A "raster" object with counts of buildings per pixel.}
}
\value{
Gridded population estimates as a "raster" object.
}
\description{
Create gridded population estimates by disaggregating population totals from polygon geometries using building footprints.
}
