% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jelly.R
\name{jelly}
\alias{jelly}
\title{Run the peanutButter::jelly shiny application}
\usage{
jelly()
}
\description{
A shiny web application to produce population estimates from building footprints using the peanut butter method.
}
