% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popRaster.R
\name{popRaster}
\alias{popRaster}
\title{popRaster}
\usage{
popRaster(
  country,
  srcdir,
  people_urb = 5,
  units_urb = 1,
  prob_urb = 0.5,
  people_rur = 5,
  prob_rur = 0.5,
  units_rur = 1
)
}
\arguments{
\item{country}{Country selection (ISO3)}

\item{srcdir}{Source directory}

\item{people_urb}{Average number of people per housing unit}

\item{units_urb}{Average number of housing units per building}

\item{prob_urb}{Probability of residential building}

\item{people_rur}{Average number of people per housing unit}

\item{prob_rur}{Probability of residential building}

\item{units_rur}{Average number of housing units per building}
}
\description{
Create population raster for sansModel
}
