% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popRaster.R
\name{popRaster}
\alias{popRaster}
\title{popRaster}
\usage{
popRaster(
  buildings_path,
  urban_path,
  people_urb = 5,
  units_urb = 1,
  residential_urb = 0.5,
  people_rur = 5,
  residential_rur = 0.5,
  units_rur = 1
)
}
\arguments{
\item{buildings_path}{Path to a raster (.tif) with building counts per pixel}

\item{urban_path}{Path to a raster (.tif) with binary map of urban areas (i.e. 0 and 1)}

\item{people_urb}{Average number of people per housing unit}

\item{units_urb}{Average number of housing units per building}

\item{residential_urb}{Probability of residential building}

\item{people_rur}{Average number of people per housing unit}

\item{residential_rur}{Probability of residential building}

\item{units_rur}{Average number of housing units per building}
}
\description{
Create population raster using the peanut butter method.
}
